refactor: AI視界描画をHUDに移動する
